Improved ID obfuscation (#1775)
* Make running kiota tests faster by turning off default handlers (such as retries)

* Remove redundant TranslateAsync calls in tests

* Fix inconsistencies in test assertions

* Assert on obfuscated links, mention sqids

* Switch from int to long, use invariant culture, use singleton codec

* Add API to remove a resource from the graph (after assembly scanning)

* Add test support for obfuscation in operations

* Add annotation to hide underlying resource ID type in OpenAPI document

* Add OpenAPI tests for ID obfuscation

+++ test/JsonApiDotNetCoreTests/IntegrationTests/IdObfuscation/DebitCardsController.cs
@@ -5,5 +5,5 @@ using Microsoft.Extensions.Logging;
 namespace JsonApiDotNetCoreTests.IntegrationTests.IdObfuscation;
 
 public sealed class DebitCardsController(
-    IJsonApiOptions options, IResourceGraph resourceGraph, ILoggerFactory loggerFactory, IResourceService<DebitCard, int> resourceService)
+    IJsonApiOptions options, IResourceGraph resourceGraph, ILoggerFactory loggerFactory, IResourceService<DebitCard, long> resourceService)
     : ObfuscatedIdentifiableController<DebitCard>(options, resourceGraph, loggerFactory, resourceService);
